requirements.txt
alembic==0.9.6
attrs==17.4.0
blinker==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
crcmod==1.7
Faker==0.8.9
Flask==0.12.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Test==0.1.6
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gevent==1.2.2
gevent-websocket==0.10.1
greenlet==0.4.12
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
jsonref==0.1
jsonschema==2.6.0
MarkupSafe==1.0
pluggy==0.6.0
py==1.5.2
pycrypto==2.6.1
PyJWT==1.5.3
PyMySQL==0.7.11
pytest==3.3.2
PyYAML==3.12
raven==6.4.0
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
Werkzeug==0.13
WTForms==2.1
Flask-Cors==3.0.7
Flask-SocketIO
